Pipelined register accesses (#25)
Register accesses can be pipelined over multiple cycles using `--option
pipelining=N` and matching that by setting the `RF_PIPELINING` parameter
to the same value when instancing `pt_axi4lite_bridge`

The pipelining parameter breaks read muxing over the N cycles given,
attempting to create a better balanced registered muxing structure to
spread X-to-1 fan-in over the N cycles.

**Explanation of using the Nth root**

For simplicity, the only control provided for pipelining determines the
number of mux-and-flop stages, the template therefore needs to determine
a close-to-optimum split of muxing across the available stages.

Take the case of 100 registers being defined, requiring a 100-to-1
fan-in spread over N stages. If we set N to three, then the optimal
fan-in configuration is:

 * 20x 5-to-1 fan-in - going from 100 to 20 unique items;
 * 4x 5-to-1 fan-in - going from 20 to 4 unique items;
 * 1x 4-to-1 fan-in - going from 4 to 1 unique items;

This achieves no one stage (with the slight exception of the final one)
becoming a timing bottleneck as the mux ratio is kept small and balanced
across the available pipelining flops.

Working this back, the underlying maths is calculating the mux ratio as
the Nth root - for example where N=3 (i.e. 3 stages of pipelining) `100
** (1/3) = 4.642` (i.e. rounding down achieves the optimal scenario).

In the code of this PR, it instead rounds up (i.e. `math.ceil(...)`)
rather than down as this achieved a better overall characteristic in
experiments.